requirements-deploy.txt
github3.py==0.9.6
PyYAML==3.12
